Prevent error in pre-save signal for Attachment
Because the new file scanning feature might delete the file outside of
Attachment class signals, add a guard to prevent an AttributeError that
would happen when trying to delete an already deleted file.

+++ forms/signals.py
@@ -1,6 +1,7 @@
 import os
 
 from django.db import models
+from django.db.models.fields.files import FieldFile
 from django.dispatch import receiver
 
 from forms.models.form import Attachment
@@ -14,16 +15,22 @@ def auto_delete_file_on_delete(sender, instance, **kwargs):
 
 
 @receiver(models.signals.pre_save, sender=Attachment)
-def auto_delete_file_on_change(sender, instance, **kwargs):
+def auto_delete_file_on_change(sender, instance: Attachment, **kwargs):
     if not instance.pk:
         return False
 
     try:
-        old_file = Attachment.objects.get(pk=instance.pk).attachment
+        old_file: FieldFile = Attachment.objects.get(pk=instance.pk).attachment
     except Attachment.DoesNotExist:
         return False
 
-    new_file = instance.attachment
+    if not old_file:
+        # FieldFile might exist, but doesn't reference a file anymore
+        # --> no need to delete anything
+        return False
+
+    new_file: FieldFile = instance.attachment
     if not old_file == new_file:
+        # Attachment was changed --> delete the previous attachment file
         if os.path.isfile(old_file.path):
             os.remove(old_file.path)
